test(app): extract common fixtures to conftest
