+++ tests/app/test_main.py
@@ -1,70 +1,22 @@
 from pathlib import Path
 from typing import List, NamedTuple, Optional
-from unittest.mock import MagicMock
 
 import pytest
 import typer
-from micropy import logger, utils
+from micropy import utils
 from micropy.app import main as main_app
 from micropy.app.main import TemplateEnum, app
-from micropy.main import MicroPy
 from micropy.project import Project
-from micropy.stubs.stubs import DeviceStub, StubManager
 from pytest_mock import MockFixture
-from typer.testing import CliRunner
-
-
-@pytest.fixture()
-def runner():
-    runner = CliRunner()
-    with runner.isolated_filesystem():
-        yield runner
-
+from tests.app.conftest import MicroPyScenario, context_mock
 
 # current working directory symbol
 CWD = object()
 
 
-class MicroPyScenario(NamedTuple):
-    project_exists: bool = True
-    has_stubs: bool = True
-
-
-@pytest.fixture()
-def micropy_obj(request: pytest.FixtureRequest, tmp_path, mocker: MockFixture) -> MicroPy:
-    mpy = mocker.MagicMock(MicroPy, autospec=True)
-    mpy.log = logger.Log.add_logger("MicroPy")
-    mpy.project = mocker.MagicMock(Project, autospec=True).return_value
-    mpy.stubs = mocker.MagicMock(StubManager, autospec=True).return_value
-    mpy.stubs.add = lambda i: i
-    if param := getattr(request, "param", MicroPyScenario()):
-        mpy.project.exists = param.project_exists
-        stubs = ["a-stub"] if param.has_stubs else []
-        mpy.stubs.__iter__.return_value = iter(stubs)
-    mocker.patch("micropy.main.MicroPy", return_value=mpy)
-    return mpy
-
-
-@pytest.fixture(params=[True, False])
-def context_mock(request: pytest.FixtureRequest, mocker: MockFixture, micropy_obj) -> MagicMock:
-    ctx = mocker.MagicMock()
-    ctx.ensure_object.return_value = micropy_obj
-    ctx.find_object.return_value = micropy_obj
-    ctx.resilient_parsing = request.param
-    ctx.obj = micropy_obj
-    return ctx
-
-
 @pytest.fixture
 def project_path(tmp_path):
     return tmp_path / "test_project"
-
-
-@pytest.fixture
-def mock_manager(mocker: MockFixture):
-    stubs = [mocker.MagicMock(spec=DeviceStub)]
-    mocker.patch.object(StubManager, "_loaded", new=stubs)
-    yield stubs
 
 
 class TemplateParamCase(NamedTuple):
